added EAIs
