+++ router/nginx.conf
@@ -5,7 +5,7 @@ http {
   server {
     listen 8000;
     listen [::]:8000;
-    server_name localhost;
+    server_name _;
 
     location / {
       proxy_pass  http://$FE_SERVER/;
